+++ src/llama_stack_client/lib/cli/models/models.py
@@ -5,8 +5,9 @@
 # the root directory of this source tree.
 
 import click
-from llama_stack_client.lib.cli.models.get import get_model
-from llama_stack_client.lib.cli.models.list import list_models
+from tabulate import tabulate
+from llama_stack_client.lib.cli.common.utils import print_table_from_response
+from typing import Optional
 
 
 @click.group()
@@ -15,6 +16,63 @@ def models():
     pass
 
 
+@click.command(name="list", help="Show available llama models at distribution endpoint")
+@click.pass_context
+def list_models(ctx):
+    client = ctx.obj["client"]
+
+    headers = ["identifier", "provider_id", "provider_resource_id", "metadata"]
+    response = client.models.list()
+    if response:
+        print_table_from_response(response, headers)
+
+
+@click.command(name="get")
+@click.argument("model_id")
+@click.pass_context
+def get_model(ctx, model_id: str):
+    """Show available llama models at distribution endpoint"""
+    client = ctx.obj["client"]
+
+    models_get_response = client.models.retrieve(identifier=model_id)
+
+    if not models_get_response:
+        click.echo(
+            f"Model {model_id} is not found at distribution endpoint. "
+            "Please ensure endpoint is serving specified model."
+        )
+        return
+
+    headers = sorted(models_get_response.__dict__.keys())
+    rows = []
+    rows.append([models_get_response.__dict__[headers[i]] for i in range(len(headers))])
+
+    click.echo(tabulate(rows, headers=headers, tablefmt="grid"))
+
+
+@click.command(name="register", help="Register a new model at distribution endpoint")
+@click.argument("model_id")
+@click.option("--provider-id", help="Provider ID for the model", default=None)
+@click.option("--provider-model-id", help="Provider's model ID", default=None)
+@click.option("--metadata", help="JSON metadata for the model", default=None)
+@click.pass_context
+def register_model(
+    ctx, model_id: str, provider_id: Optional[str], provider_model_id: Optional[str], metadata: Optional[str]
+):
+    """Register a new model at distribution endpoint"""
+    client = ctx.obj["client"]
+
+    try:
+        response = client.models.register(
+            model_id=model_id, provider_id=provider_id, provider_model_id=provider_model_id, metadata=metadata
+        )
+        if response:
+            click.echo(f"Successfully registered model {model_id}")
+    except Exception as e:
+        click.echo(f"Failed to register model: {str(e)}")
+
+
 # Register subcommands
 models.add_command(list_models)
 models.add_command(get_model)
+models.add_command(register_model)
